MAJOR CODE REWORK/SPLIT - add assign permission command - npm update - fix text
