+++ .idea/jsLibraryMappings.xml
@@ -1,6 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="JavaScriptLibraryMappings">
+    <file url="file://$PROJECT_DIR$" libraries="{discord_bot_firewall/node_modules}" />
+    <file url="PROJECT" libraries="{discord_bot_firewall/node_modules}" />
+    <includedPredefinedLibrary name="HTTP Response Handler" />
     <includedPredefinedLibrary name="Node.js Core" />
   </component>
 </project>
