+++ Dockerfile
@@ -14,6 +14,6 @@ COPY . .
 ARG timezone=Etc/UTC
 ARG cronMinHour="0 0"
 RUN ln -snf /usr/share/zoneinfo/$timezone /etc/localtime && echo $timezone > /etc/timezone && chmod 777 public/exports && chmod 4755 /usr/sbin/cron && echo "$cronMinHour * * * root rm -f /usr/src/app/public/exports/*" >> /etc/cron.d/exportclean && npm install
-EXPOSE 8080
+EXPOSE 3000
 USER 1001
 CMD cron && cd /var/exports/ && python3 -m http.server & node bin/www
