+++ package.json
@@ -21,7 +21,7 @@
     "capitalize": "^2.0.3",
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
-    "discord.js": "^12.1.1",
+    "discord.js": "^12.2.0",
     "express": "~4.16.1",
     "http-build-query": "^0.7.0",
     "http-errors": "~1.6.3",
@@ -33,8 +33,8 @@
   "devDependencies": {
     "axios": "^0.19.2",
     "dotenv": "^8.2.0",
-    "nodemon": "^2.0.2",
+    "nodemon": "^2.0.3",
     "prettier": "2.0.2",
-    "shelljs": "^0.8.3"
+    "shelljs": "^0.8.4"
   }
 }
